Refactoring, Observables added, minor changes in hole game

+++ Roboti/app/src/main/java/cz/muni/fi/umimecesky/activity/RaceActivity.kt
@@ -90,7 +90,7 @@ class RaceActivity : BaseAbstractActivity() {
         button.postDelayed({
             enableButtons()
             setNewRandomWord()
-        }, RACE_NEW_WORD_DELAY_MS.toLong())
+        }, RACE_NEW_WORD_DELAY_MS)
     }
 
     override fun chosenWrong(button: Button) {
