+++ Roboti/app/src/main/java/cz/muni/fi/umimecesky/logic/BotLogicQuick.kt
@@ -5,7 +5,7 @@ import cz.muni.fi.umimecesky.pojo.RaceConcept
 
 class BotLogicQuick(private val concept: RaceConcept) : RobotLogic {
 
-    override fun millisecondsPerSolution(): Int = 3000 - (1800 * concept.levelProgress()).toInt()
+    override fun millisecondsPerSolution(): Long = 3000L - (1800L * concept.levelProgress()).toLong()
 
     override fun correctnessRatio(): Double =
             0.65 + roundBy2places(0.3 * concept.levelProgress())
